+++ Race-OOP/MagicCarpet.cs
@@ -10,7 +10,7 @@ namespace Race_OOP
         {
         }
 
-        protected override decimal GetReducedDistance(int distance)
+        protected override double GetReducedDistance(int distance)
         {
             if (distance < 1000)
             {
@@ -18,15 +18,15 @@ namespace Race_OOP
             }
             else if (distance < 5000)
             {
-                return (decimal) (distance * 0.97);
+                return (double) (distance * 0.97);
             }
             else if (distance < 10000)
             {
-                return (decimal) (distance * 0.9);
+                return (double) (distance * 0.9);
             }
             else
             {
-                return (decimal) (distance * 0.95);
+                return (double) (distance * 0.95);
             }
         }
     }
